chore: stepper with routing
