+++ packages/apps/dev-wallet/src/constants/color-list.ts
@@ -1,0 +1,7 @@
+export const PROFILE_COLOR_LIST = [
+  '#42CEA4',
+  '#42BDCE',
+  '#4269CE',
+  '#B242CE',
+  '#CEA742',
+];
